Return nil stubs for unimplemented rumor tracking
This seems to be what was planned to address the problem in GearHead 1
where job and mission hints would be lost amidst a sea of throwaway
personality hints and madlibs.  It's not clear how it should work
exactly: how does the game tell that you've seen a rumor, or when they
should be deleted, or avoid duplication when two NPCs tell the same
rumor with different words?

I solved this problem in my personal copy of GH1 by filtering the job
and mission hints into the message area (not published).