Filter for trigger-eligible gears at a location
i.e. Standing on stairs and pressing '>' just uses the stairs instead of
bringing up a menu with the PC and the stairs and making the player
choose the stairs.

This mirrors the design present in GH2, but integrates it with GHAR's
Lua system, so instead of checking for a 'USE' SAtt trigger script it
finds the gear in Lua-land and sees if it has a mapping for 'USE',
presumably the trigger handler function.

I had to comment out the "Using something!" stub, otherwise this
wouldn't have fixed the stairs issue.